only two classes

+++ src/trees.py
@@ -77,9 +77,10 @@ class LeafTreebankNode(TreebankNode):
 class ParseNode(object):
     pass
 
-
 def create_internal_parse_node(label, children):
-    if len(children) == 1 and isinstance(children[0], InternalParseNode):
+    if len(children) == 0:
+        return None
+    elif len(children) == 1 and isinstance(children[0], InternalParseNode):
         child = children[0]
         return create_internal_parse_node(label + child.label, child.children)
     else:
@@ -104,47 +105,23 @@ class InternalParseNode(ParseNode):
         self.right = children[-1].right
 
         self.leaves = [leaf for child in self.children for leaf in child.leaves]
-        self.tree_bank_version = None
-
-
-    def clean_up_punctuation(self):
-        new_children = []
-        for child in self.children:
-            child = child.clean_up_punctuation()
-            if isinstance(child, LeafParseNode):
-                new_children.append(child)
-            else:
-                child_children = child.children
-                while len(child_children) > 0 and isinstance(child_children[0], LeafParseNode) and child_children[0].tag in deletable_tags:
-                    new_children.append(child_children[0])
-                    child_children = child_children[1:]
-
-                more_children = []
-                while len(child_children) > 0 and isinstance(child_children[-1], LeafParseNode) and child_children[-1].tag in deletable_tags:
-                    more_children.append(child_children[-1])
-                    child_children = child_children[:-1]
-
-                if len(child_children) > 0:
-                    child = create_internal_parse_node(child.label, child_children)
-                    new_children.append(child)
-                new_children.extend(more_children)
-        return create_internal_parse_node(self.label, new_children)
+        self._delete_punctuation = None
+        self._tree_bank = None
 
     def delete_punctuation(self):
-        new_children = []
-        for child in self.children:
-            if isinstance(child, InternalParseNode):
-                new_child = child.delete_punctuation()
-                if new_child is not None:
-                    new_children.append(new_child)
-            else:
-                assert isinstance(child, LeafParseNode)
-                if child.tag not in deletable_tags:
-                    new_children.append(child)
-        if len(new_children) == 0:
-            return None
-        else:
-            return create_internal_parse_node(self.label, new_children)
+        if self._delete_punctuation is None:
+            new_children = []
+            for child in self.children:
+                if isinstance(child, InternalParseNode):
+                    new_child = child.delete_punctuation()
+                    if new_child is not None:
+                        new_children.append(new_child)
+                else:
+                    assert isinstance(child, LeafParseNode)
+                    if child.tag not in deletable_tags:
+                        new_children.append(child)
+            self._delete_punctuation = create_internal_parse_node(self.label, new_children)
+        return self._delete_punctuation
 
     def reset(self, left):
         self.left = left
@@ -158,13 +135,13 @@ class InternalParseNode(ParseNode):
         self.right = cur_left
 
     def convert(self):
-        if self.tree_bank_version is None:
+        if self._tree_bank is None:
             children = [child.convert() for child in self.children]
             tree = InternalTreebankNode(self.label[-1], children)
             for sublabel in reversed(self.label[:-1]):
                 tree = InternalTreebankNode(sublabel, [tree])
-            self.tree_bank_version = tree
-        return self.tree_bank_version
+            self._tree_bank = tree
+        return self._tree_bank
 
     def enclosing(self, left, right):
         assert self.left <= left < right <= self.right, (self.left, left, right, self.right)
@@ -203,9 +180,6 @@ class LeafParseNode(ParseNode):
         self.word = word
         self.leaves = [self]
 
-    def clean_up_punctuation(self):
-        return self
-
     def reset(self, left):
         self.left = left
         self.rigth = left + 1
@@ -214,7 +188,10 @@ class LeafParseNode(ParseNode):
         return LeafTreebankNode(self.tag, self.word)
 
     def delete_punctuation(self):
-        return self
+        if self.tag in deletable_tags:
+            return None
+        else:
+            return self
 
 
 def load_trees(path, strip_top=True):
